+++ test/websocket/doc_snippets.cpp
@@ -217,7 +217,7 @@ boost::asio::ip::tcp::socket sock{ios};
 } // fxx()
 
 // workaround for https://github.com/chriskohlhoff/asio/issues/112
-#ifdef _MSC_VER
+#ifdef BOOST_MSVC
 //[ws_snippet_21
 void echo(stream<boost::asio::ip::tcp::socket>& ws,
     multi_buffer& buffer, boost::asio::yield_context yield)